Dominand_writing_direction_exercise
This exercise relies on a SCRIPT DATA SET, which is not included in this code. Therefore, the code above may not function correctly outside of the Eloquent JavaScript 4th Edition console. Additionally, there is a known bug in the solution where, if the text contains the same quantity of words with different writing directions, the function might incorrectly return the direction of the last word as the dominant one. This should be taken into account when testing the function, as it may lead to inaccurate results in cases of mixed scripts.