Pyramid FCC Project
This exercise is a small project from the new JavaScript Algorithms and Data Structures (Beta) Certification on FreeCodeCamp. It focuses on generating a simple text-based pyramid pattern using loops and string manipulation techniques in JavaScript. The project helps to reinforce fundamental programming concepts such as loops, functions, and string operations.